reworked external_panel.py
