+++ src/gtnh/gui/external/external_panel.py
@@ -1,8 +1,17 @@
-from tkinter import LabelFrame
-from typing import Any, Dict, List, Optional
+import asyncio
+from tkinter import LabelFrame, Toplevel
+from tkinter.messagebox import showerror
+from typing import Any, Callable, Coroutine, Dict, List, Optional
 
-from gtnh.gui.external.external_modlist import ExternalModList
+from gtnh.defs import Position
+from gtnh.gui.external.mod_adder_window import ModAdderWindow
+from gtnh.gui.lib.button import CustomButton
+from gtnh.gui.lib.custom_widget import CustomWidget
+from gtnh.gui.lib.listbox import CustomListbox
 from gtnh.gui.mod_info.mod_info_widget import ModInfoWidget
+from gtnh.models.gtnh_version import GTNHVersion
+from gtnh.models.mod_info import ExternalModInfo
+from gtnh.modpack_manager import GTNHModpackManager
 
 
 class ExternalPanel(LabelFrame):
@@ -24,8 +33,6 @@ class ExternalPanel(LabelFrame):
         self.xpadding: int = 0
         LabelFrame.__init__(self, master, text=frame_name, **kwargs)
 
-        self.width: Optional[int] = width
-
         mod_info_callbacks: Dict[str, Any] = {
             "set_mod_version": callbacks["set_external_mod_version"],
             "set_mod_side": callbacks["set_external_mod_side"],
@@ -34,28 +41,47 @@ class ExternalPanel(LabelFrame):
             self, frame_name="External mod info", callbacks=mod_info_callbacks
         )
 
-        external_mod_list_callbacks: Dict[str, Any] = {
-            "get_gtnh": callbacks["get_gtnh"],
-            "get_external_mods": callbacks["get_external_mods"],
-            "mod_info": self.mod_info_frame.populate_data,
-            "add_mod_in_memory": callbacks["add_mod_in_memory"],
-            "del_mod_in_memory": callbacks["del_mod_in_memory"],
-            "freeze": callbacks["freeze"],
-            "refresh_external_mods": callbacks["refresh_external_mods"],
-        }
+        # start
+        self.get_gtnh_callback: Callable[[], Coroutine[Any, Any, GTNHModpackManager]] = callbacks["get_gtnh"]
+        self.get_external_mods_callback: Callable[[], Dict[str, str]] = callbacks["get_external_mods"]
+        self.toggle_freeze: Callable[[], None] = callbacks["freeze"]
+        self.mod_info_callback: Callable[[Any], None] = self.mod_info_frame.populate_data
+        self.add_mod_to_memory: Callable[[str, str], None] = callbacks["add_mod_in_memory"]
+        self.del_mod_from_memory: Callable[[str], None] = callbacks["del_mod_in_memory"]
+        self.refresh_external_modlist: Callable[[], None] = callbacks["refresh_external_mods"]
 
-        self.external_mod_list: ExternalModList = ExternalModList(
-            self, frame_name="External mod list", callbacks=external_mod_list_callbacks
+        self.listbox: CustomListbox = CustomListbox(
+            self,
+            label_text="External mods:",
+            exportselection=False,
+            on_selection=lambda event: asyncio.ensure_future(self.on_listbox_click(event)),
+            display_horizontal_scrollbar=True,
         )
 
-        if self.width is None:
-            self.width = self.external_mod_list.get_width()
+        self.btn_add: CustomButton = CustomButton(
+            self, text="Add new mod", command=lambda: asyncio.ensure_future(self.add_external_mod())
+        )
+
+        self.btn_add_version: CustomButton = CustomButton(
+            self, text="Add new version to highlighted", command=lambda: asyncio.ensure_future(self.add_new_version())
+        )
+
+        self.btn_rem: CustomButton = CustomButton(
+            self, text="Delete highlighted", command=lambda: asyncio.ensure_future(self.del_external_mod())
+        )
+
+        self.widgets: List[CustomWidget] = [self.btn_add, self.btn_rem, self.btn_add_version, self.listbox]
+        self.width: int = (
+            width if width is not None else max([widget.get_description_size() for widget in self.widgets])
+        )
+
+        if width is None:
             self.mod_info_frame.set_width(self.width)
             self.update_widget()
 
         else:
             self.mod_info_frame.set_width(self.width)
-            self.external_mod_list.set_width(self.width)
+            self.update_widget()
 
     def configure_widgets(self) -> None:
         """
@@ -64,7 +90,8 @@ class ExternalPanel(LabelFrame):
         :return: None
         """
         self.mod_info_frame.configure_widgets()
-        self.external_mod_list.configure_widgets()
+        for widget in self.widgets:
+            widget.configure(width=self.width)
 
     def set_width(self, width: int) -> None:
         """
@@ -75,7 +102,7 @@ class ExternalPanel(LabelFrame):
         """
         self.width = width
         self.mod_info_frame.set_width(self.width)
-        self.external_mod_list.set_width(self.width)
+        self.configure_widgets()
 
     def get_width(self) -> int:
         """
@@ -83,7 +110,6 @@ class ExternalPanel(LabelFrame):
 
         :return: the width in character sizes of the normalised widgets
         """
-        assert self.width  # can't be None because how it's defined in the constructor
         return self.width
 
     def update_widget(self) -> None:
@@ -96,7 +122,6 @@ class ExternalPanel(LabelFrame):
         self.configure_widgets()
         self.show()
 
-        self.external_mod_list.update_widget()
         self.mod_info_frame.update_widget()
 
     def hide(self) -> None:
@@ -104,10 +129,10 @@ class ExternalPanel(LabelFrame):
         Method to hide the widget and all its childs
         :return None:
         """
-        self.external_mod_list.grid_forget()
+        for widget in self.widgets:
+            widget.grid_forget()
         self.mod_info_frame.grid_forget()
 
-        self.external_mod_list.hide()
         self.mod_info_frame.hide()
 
         self.update_idletasks()
@@ -120,8 +145,8 @@ class ExternalPanel(LabelFrame):
         """
         x: int = 0
         y: int = 0
-        rows: int = 2
-        columns: int = 1
+        rows: int = 1
+        columns: int = 2
 
         for i in range(rows):
             self.rowconfigure(i, weight=1, pad=self.xpadding)
@@ -129,10 +154,13 @@ class ExternalPanel(LabelFrame):
         for i in range(columns):
             self.columnconfigure(i, weight=1, pad=self.ypadding)
 
-        self.external_mod_list.grid(row=x, column=y)
-        self.mod_info_frame.grid(row=x + 1, column=y)
+        self.listbox.grid(row=x, column=y, columnspan=2, sticky=Position.HORIZONTAL)
+        self.btn_add.grid(row=x + 1, column=y, sticky=Position.HORIZONTAL)
+        self.btn_rem.grid(row=x + 1, column=y + 1, sticky=Position.HORIZONTAL)
+        self.btn_add_version.grid(row=x + 2, column=y, sticky=Position.HORIZONTAL)
 
-        self.external_mod_list.show()
+        self.mod_info_frame.grid(row=x + 3, column=y, columnspan=2)
+
         self.mod_info_frame.show()
 
         self.update_idletasks()
@@ -145,4 +173,129 @@ class ExternalPanel(LabelFrame):
         :return: None
         """
         mod_list: List[str] = data["external_mod_list"]
-        self.external_mod_list.populate_data(mod_list)
+        self.listbox.set_values(sorted(mod_list))
+
+    async def on_listbox_click(self, _: Any) -> None:
+        """
+        Callback used when the user clicks on the external mods' listbox.
+
+        :param _: the tkinter event passed by the tkinter in the Callback (unused)
+        :return: None
+        """
+        if self.listbox.has_selection():
+            index: int = self.listbox.get()
+            gtnh: GTNHModpackManager = await self.get_gtnh_callback()
+            mod_info: ExternalModInfo = gtnh.assets.get_external_mod(self.listbox.get_value_at_index(index))
+            name: str = mod_info.name
+            mod_versions: list[GTNHVersion] = mod_info.versions
+            latest_version: Optional[GTNHVersion] = mod_info.get_latest_version()
+            assert latest_version
+            external_mods: Dict[str, str] = self.get_external_mods_callback()
+            current_version: str = external_mods[name] if name in external_mods else latest_version.version_tag
+
+            _license: str = mod_info.license or "No license detected"
+            side: str = mod_info.side
+
+            data = {
+                "name": name,
+                "versions": [version.version_tag for version in mod_versions],
+                "current_version": current_version,
+                "license": _license,
+                "side": side,
+            }
+            self.mod_info_callback(data)
+
+    async def add_external_mod(self) -> None:
+        """
+        Method called when the button to add an external mod is pressed.
+
+        :return: None
+        """
+
+        # showerror("Feature not yet implemented", "The addition of external mods to the assets is not yet implemented.")
+        # don't forget to use self.add_mod_in_memory when implementing this
+        self.toggle_freeze()
+        top_level: Toplevel = Toplevel(self)
+
+        def close(event: Any = None) -> None:
+            """
+            Method called when toplevel is destroyed.
+
+            :return: None
+            """
+            if event.widget is top_level:
+                self.toggle_freeze()
+                asyncio.ensure_future(self.refresh_external_modlist())  # type: ignore
+
+        top_level.bind("<Destroy>", close)
+        callbacks = {
+            "get_gtnh": self.get_gtnh_callback,
+            "add_mod_in_memory": self.add_mod_to_memory,
+            "del_mod_from_memory": self.del_mod_from_memory,
+        }
+        mod_addition_frame: ModAdderWindow = ModAdderWindow(top_level, "external mod adder", callbacks=callbacks)
+        mod_addition_frame.grid()
+        mod_addition_frame.update_widget()
+        top_level.title("External mod addition")
+
+    async def del_external_mod(self) -> None:
+        """
+        Method called when the button to delete the highlighted external mod is pressed.
+
+        :return: None
+        """
+        # showerror("Feature not yet implemented", "The removal of external mods from assets is not yet implemented.")
+        # don't forget to use self.del_mod_from_memory when implementing this
+        if self.listbox.has_selection():
+            index: int = self.listbox.get()
+            mod_name: str = self.listbox.get_value_at_index(index)
+            gtnh: GTNHModpackManager = await self.get_gtnh_callback()
+            self.listbox.del_value_at_index(index)
+            await gtnh.delete_external_mod(mod_name)
+        else:
+            showerror(
+                "No curseforge mod selected",
+                "In order to add a new version to a curseforge mod, you must select one first",
+            )
+            return
+
+    async def add_new_version(self) -> None:
+        """
+        Method called when the button to add a new version to an external mod is pressed.
+
+        :return: None
+        """
+        if self.listbox.has_selection():
+            index: int = self.listbox.get()
+            mod_name: str = self.listbox.get_value_at_index(index)
+            self.toggle_freeze()
+            top_level: Toplevel = Toplevel(self)
+
+            def close(event: Any = None) -> None:
+                """
+                Method called when toplevel is destroyed.
+
+                :return: None
+                """
+                if event.widget is top_level:
+                    self.toggle_freeze()
+                    asyncio.ensure_future(self.refresh_external_modlist())  # type: ignore
+
+            top_level.bind("<Destroy>", close)
+            callbacks = {
+                "get_gtnh": self.get_gtnh_callback,
+                "add_mod_in_memory": self.add_mod_to_memory,
+                "del_mod_from_memory": self.del_mod_from_memory,
+            }
+            mod_addition_frame: ModAdderWindow = ModAdderWindow(
+                top_level, "external version adder", callbacks=callbacks, mod_name=mod_name
+            )
+            mod_addition_frame.grid()
+            mod_addition_frame.update_widget()
+            top_level.title("External mod addition")
+        else:
+            showerror(
+                "No curseforge mod selected",
+                "In order to add a new version to a curseforge mod, you must select one first",
+            )
+            return
